Activity_Execution/requirements.txt
streamlit
opencv-python
ultralytics
torch
torchvision
mediapipe
Pillow
lap